+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `auto_increment` to `receiptComponents` type
-- `startingAutoIncrement` to `externalIdGeneration` configuration
+- `sequentialNumber` to `receiptComponents` type
+- `startingSequentialNumber` to `externalIdGeneration` configuration
 
 ## [9.1.0] - 2024-10-09